ZScript highlighter re-work Added keywords back after the work has been done, while bringing the whole thing up to date keyword-wise.
Note: due to a character limit per keywords group in Notepad++, the methods/functions
keywords had to be split between two keywords groups: methods and functions occupy
Keywords #4, and ACS action specials occupy Keywords #5. Constants were pushed to
Keywords #6, as a result.

If the need arises in the future, the keywords in Keywords #4 and #5 may be split evenly
between the two groups of keywords, as opposed to keeping Keywords #5 exclusive to action
specials. So always treat those two keywords groups as the same group in terms of styling.